# src/swagger/swagger.json
{
	"basePath": "/api/tnm/v1",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"GridCurrents": {
			"properties": {},
			"type": "object",
			"additionalProperties": false
		},
		"Card": {
			"properties": {
				"token": {
					"type": "string",
					"description": "RFID"
				},
				"status": {
					"type": "string",
					"enum": [
						"ACCEPTED",
						"BLOCKED",
						"EXPIRED",
						"INVALID",
						"UNKNOWN"
					]
				},
				"expirationDate": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"token",
				"status"
			],
			"type": "object",
			"additionalProperties": false
		},
...
	},
	"info": {
		"title": "ncpa-api",
		"version": "1.0.0",
		"description": "Reference API for NCPA",
		"license": {
			"name": "UNLICENSED"
		}
	},
	"paths": {
		"/auth": {
			"get": {
				"operationId": "ValidateAuthToken",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"properties": {
								"accessToken": {
									"type": "string"
								},
								"publicKey": {
									"type": "string"
								}
							},
							"required": [
								"publicKey",
								"accessToken"
							],
							"type": "object"
						}
					}
				},
				"description": "expects token as url encoded ciphered jwt token like so:\n```js\nconst payload =\n{\n\"iss\": \"TNM Auth server\",\n\"sub\": \"{cp-uuid}\",\n\"aud\": \"operator\" | \"installer\",\n\"iat\": {unix time}},\n\"wifi\": {\n\"ssid\": \"my-ssid\",\n\"password\": \"strong-wifi-password\",\n\"type\": \"wpa2\",\n\"hidden\": true\n};\n\nconst jwtToken = jwt.sign(payload, privateKey, {\nalgorithm: 'RS256'\n});\n\nconst token = encodeURIComponent(encrypt(jwtToken));\n\n\n```\nreturns an access token and its accompanying public key for signature validation\n\n```json\n{\naccessToken: \"eyJhbGciOiJSUzI1NiI...\",\npublicKey: \"-----BEGIN CERTIFI...\"\n}\n```",
				"tags": [
					"public"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "clientId",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/authentication-mode": {
			"get": {
				"operationId": "GetAuthenticationMode",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"PLUGNCHARGE",
								"WHITELIST"
							]
						}
					}
				},
				"description": "jwt scopes: `operator`, `installer`",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "SetAuthenticationMode",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"PLUGNCHARGE",
								"WHITELIST"
							]
						}
					}
				},
				"description": "jwt scopes: `operator`, `installer`",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "authenticationMode",
						"required": true,
						"type": "string",
						"enum": [
							"PLUGNCHARGE",
							"WHITELIST"
						]
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/config/socket-lock-mode": {
			"get": {
				"operationId": "GetSocketLockMode",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"TRANSACTION",
								"LOCKED",
								"UNLOCKED"
							]
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "SetSocketLockMode",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"TRANSACTION",
								"LOCKED",
								"UNLOCKED"
							]
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "socketLockMode",
						"required": true,
						"type": "string",
						"enum": [
							"TRANSACTION",
							"LOCKED",
							"UNLOCKED"
						]
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/config/grid-max-current": {
			"get": {
				"operationId": "GetGridMaxCurrent",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "SetGridMaxCurrent",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "maxCurrent",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/config/load-shedding": {
			"get": {
				"operationId": "GetLoadShedding",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"NO",
								"P1",
								"XEMEX_BLACK",
								"XEMEX_9600"
							]
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "SetLoadShedding",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"NO",
								"P1",
								"XEMEX_BLACK",
								"XEMEX_9600"
							]
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "loadShedding",
						"required": true,
						"type": "string",
						"enum": [
							"NO",
							"P1",
							"XEMEX_BLACK",
							"XEMEX_9600"
						]
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/config/charge-station-max-current": {
			"get": {
				"operationId": "GetChargeStationMaxCurrent",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "SetChargeStationMaxCurrent",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "chargeStationMaxCurrent",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/config/on-off-peak": {
			"get": {
				"operationId": "GetOnOffPeak",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"ON_PEAK",
								"OFF_PEAK",
								"TRANSACTION_PEAK"
							]
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "SetOnOffPeak",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"ON_PEAK",
								"OFF_PEAK",
								"TRANSACTION_PEAK"
							]
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "onOffPeak",
						"required": true,
						"type": "string",
						"enum": [
							"ON_PEAK",
							"OFF_PEAK",
							"TRANSACTION_PEAK"
						]
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/config/grid-currents": {
			"get": {
				"operationId": "GetGridCurrents",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/GridCurrents"
						}
					}
				},
				"description": "jwt scopes: `installer`",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "SetGridCurrents",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/GridCurrents"
						}
					}
				},
				"description": "jwt scopes: `installer` \n\nmin 1 item, max 3",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "gridCurrents",
						"required": true,
						"type": "array",
						"items": {
							"type": "number"
						},
						"minItems": 1,
						"maxItems": 3
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/config/reboot": {
			"post": {
				"operationId": "Reboot",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "jwt scopes: `installer` \n\nReboots CP",
				"tags": [
					"installer"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
		"/whitelist/cards": {
			"get": {
				"operationId": "GetCards",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/Card"
							},
							"type": "array"
						}
					}
				},
				"description": "jwt scopes: `operator`, `installer`",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/whitelist/cards/{token}": {
			"get": {
				"operationId": "GetCard",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Card"
						}
					}
				},
				"description": "jwt scopes: `operator`, `installer`",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "token",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "DeleteCard",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Card"
						}
					}
				},
				"description": "jwt scopes: `operator`, `installer`",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "token",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
		"/charging/start": {
			"post": {
				"operationId": "StartCharging",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "jwt scopes: `operator` `",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "tag",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/charging/stop": {
			"post": {
				"operationId": "StopCharging",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "jwt scopes: `operator`",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "tag",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/charging/unlock": {
			"post": {
				"operationId": "Unlock",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "jwt scopes: `operator`",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "access_token",
			"in": "query"
		},
		"jwtAuth": {
			"type": "apiKey",
			"in": "header",
			"name": "authorization"
		}
	}
}